requirements.txt
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
PyJWT==2.8.0
typing_extensions==4.10.0
